fix: refactor layout and page components for improved type safety and structure
