+++ src/components/index.ts
@@ -1,0 +1,4 @@
+export * from "./common";
+export * from "./home";
+export * from "./icons";
+export * from "./layout";
